Add files via upload
I have implemented and practice some basic text processing techniques in NLP for theUrdu language. The Urdu language is written in a different style as compared to English, making segmentation a challenging task. I have done word and sentence segmentation in this project and finally find the accuracy of final output.